sys-cluster/openmpi: don't maintain multiple versions
Unless there's some very specific reason to do so.

Generally, let the git history of this ebuilds preserve old versions;
if need to build and old version, then extract a version from git into
a new file. No need to carry many ebuilds that are nearly the same
because making a generic change to the ebuild means having to
change multiple files.