sys-cluster/prrte: build ALPS plugins as DSO
The plugins that themselves depend on an external .so library,
must not be linked in statically, because that external dep
becomes a dep of all mpi binaries, including mpicc, etc.

In latest master, this is the default already, but we didn't
bump the snapshot yet.

See https://github.com/open-mpi/ompi/pull/8800