dev-python/dolfin: remove unnecessary patch for PETSc include path
This was resolved by rebuilding sci-mathematics/dolfin, on first
build it did not find PETSc. TODO: check that this doesn't happen
on every build-from-scratch of the prefix (perhaps the PETSC_DIR
env vars are not loaded? but this isn't a problem with any other
packages, though, so not sure).

Also, the fix to sci-mathematics/dolfin to look for lowercase .pc
files for PETSc and SLEPc committed recently to this ebuilds repo is
related.